refactor: move components to individual files
